Refactor Tests by Functionality
